pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "syswatch"
version = "0.1.0"
description = "CLI tool that collects system metrics and uses an LLM to narrate diagnostics"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "psutil>=5.9",
    "python-dotenv>=1.0",
    "anthropic>=0.25",
    "httpx>=0.27",
]

[project.scripts]
syswatch = "syswatch.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
